// packages/uimodel-composer/src/generated/BaseStyle.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { Style } from './Style';
import type { Expression } from './Expression';

/**
 * BaseStyle
 * @generated
 */
export interface BaseStyle extends Style {
  extends?: BaseStyle;
  css?: string;
  vueComponent?: string;
  visibilityCondition?: Expression;

}