// packages/uimodel-composer/src/generated/LayoutStyle.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { BaseStyle } from './BaseStyle';
import type { LayoutType } from './LayoutType';

/**
 * LayoutStyle
 * @generated
 */
export interface LayoutStyle extends BaseStyle {
  layout?: LayoutType;
  order?: number;

}