// packages/uimodel-composer/src/generated/WidgetStyle.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EStructuralFeature, BasicEObject, EClass } from '@emfts/core';
import type { BaseStyle } from './BaseStyle';
import type { WidgetType } from './WidgetType';

/**
 * WidgetStyle
 * @generated
 */
export interface WidgetStyle extends BaseStyle {
  feature?: EStructuralFeature;
  widgetType?: WidgetType;
  label?: string;
  readOnly?: boolean;
  order?: number;

}